# static/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: CLN REST API
  version: "1.5"

# ---------------- Security Schemes ----------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SEO:
      type: object
      properties:
        page:
          type: string
          example: home
        title:
          type: string
          example: CLN | Home
        description:
          type: string
          example: CLN Cambodia provides international and domestic logistics services.
        keywords:
          type: string
          example: CLN Cambodia, logistics, transportation, sea freight, air freight
        ogTitle:
          type: string
          example: CLN Cambodia
        ogDescription:
          type: string
          example: Offering international and domestic logistics services with 20 years of experience.
        ogImage:
          type: string
          example: https://cln-tan.vercel.app/og-home.png
        url:
          type: string
          example: https://cln-tan.vercel.app/home
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Page not found
paths:
...
  /api/v1/docs/seo:
    get:
      summary: Get all SEO entries
      description: Retrieve a list of all SEO entries from the database
      tags:
        - SEO
      responses:
        '200':
          description: List of SEO entries retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SEO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create a new SEO entry
      tags:
        - SEO
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SEO'
      responses:
        '201':
          description: SEO created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: SEO created
        '400':
          description: SEO for this page already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/docs/seo/{page}:
    get:
      summary: Get SEO data for a page
      tags:
        - SEO
      parameters:
        - in: path
          name: page
          schema:
            type: string
          required: true
          description: Page identifier, e.g., "services"
        - in: query
          name: tab
          schema:
            type: string
          required: false
          description: Optional tab or subpage, e.g., "custom" or "land"
      responses:
        '200':
          description: SEO data retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SEO'
        '404':
          description: Page not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update SEO data for a page
      tags:
        - SEO
      parameters:
        - in: path
          name: page
          schema:
            type: string
          required: true
          description: Page identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SEO'
      responses:
        '200':
          description: SEO updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: SEO updated
        '404':
          description: Page not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete SEO entry for a page
      tags:
        - SEO
      parameters:
        - in: path
          name: page
          schema:
            type: string
          required: true
          description: Page identifier
      responses:
        '200':
          description: SEO deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: SEO deleted
        '404':
          description: Page not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...